rename gfm_client.h to gfm_client_internal.h
