+++ Makefile.am
@@ -2,7 +2,7 @@ lib_LTLIBRARIES = libglobus_gridftp_server_gfarm.la
 libglobus_gridftp_server_gfarm_la_SOURCES = globus_gridftp_server_gfarm.c
 EXTRA_DIST = RELNOTES LICENSE README.md
 
-dist_noinst_HEADERS = gfm_client.h
+dist_noinst_HEADERS = gfm_client_internal.h
 
 AM_CPPFLAGS = -Wall @GFARM_CFLAGS@ @GRIDFTP_SERVER_CFLAGS@
 AM_LDFLAGS = @GFARM_LIBS@ @GRIDFTP_SERVER_LIBS@
