+++ gfm_client_internal.h
